refac: Retire IActorRegistry (#678)
IActorRegistry is proving to be slightly annoying to handle in the
master data module introduction. As it only serves to make life easier
when running in dev mode, the component is removed for now.
After the master data module introduction is complete, IActorRegistry,
or more likely, the IActorRepository, might once again provide the
feature of automatic actor creation in dev mode.