fix: Use PM event in load test
Since t001 has been updated to use the PM version of the calculation
flow, EDI will no longer act upon the event from wholesale. This had
the side effect of making the load test do nothing, as it was based on
reception of an event from wholesale.

The load test has been updated to instead publish the PM event. While
we could have looked at the feature flag and used that to determine
which event to publish, we are so close to fully changing to the PM
flow that it would be a wasted effort.